refactor: #14 ProductController 이름 변경 및 API 정리
